requirements.txt
agno==1.8.4
requests
python-dotenv
openai
fastapi
uvicorn
python-multipart
numpy
faiss-cpu
beautifulsoup4
selenium
python-docx
striprtf
PyMuPDF
pillow
easyocr
torch
torchvision
pandas